requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.6.0
yfinance>=0.2.0
pyyaml>=6.0
pyarrow>=12.0.0
joblib>=1.2.0
